requirements.txt
# Agentic Long-Term Memory - Requirements
# For Windows users experiencing compilation issues, use: python install.py

# Core dependencies (should install on all platforms)
gradio==5.23.1
ipython==8.12.3
openai==1.69.0
pandas==2.2.3
pydantic
pyprojroot==0.3.0
python-dotenv==1.1.0
PyYAML==6.0.2
tiktoken==0.9.0
typing_extensions==4.13.0
langchain==0.3.20
langgraph==0.3.5
langmem==0.0.16
ipykernel==6.29.5
SQLAlchemy
langchain_community

# Vector Database (may require C++ compiler on Windows)
# If installation fails, run: python install.py
chromadb==0.6.3

# Alternative vector databases (uncomment if chromadb fails):
# faiss-cpu==1.9.0
# weaviate-client==4.9.5